move to vnode dir
