+++ src/common/src/hash/key.rs
@@ -36,10 +36,11 @@ use crate::array::{
     StructRef,
 };
 use crate::collection::estimate_size::EstimateSize;
+use crate::hash::vnode::VirtualNode;
 use crate::row::Row;
 use crate::types::{
     DataType, Decimal, IntervalUnit, NaiveDateTimeWrapper, NaiveDateWrapper, NaiveTimeWrapper,
-    OrderedF32, OrderedF64, ScalarRef, ToOwnedDatum, VirtualNode,
+    OrderedF32, OrderedF64, ScalarRef, ToOwnedDatum,
 };
 use crate::util::hash_util::Crc32FastBuilder;
 use crate::util::value_encoding::{deserialize_datum, serialize_datum};
