+++ src/common/src/hash/mod.rs
@@ -16,3 +16,5 @@ mod key;
 pub use key::*;
 mod dispatcher;
 pub use dispatcher::HashKeyDispatcher;
+mod vnode;
+pub use vnode::*;
