+++ src/common/src/util/worker_util.rs
@@ -16,7 +16,7 @@ use std::collections::HashMap;
 
 use risingwave_pb::common::WorkerNode;
 
-use crate::types::ParallelUnitId;
+use crate::hash::vnode::ParallelUnitId;
 
 pub fn get_pu_to_worker_mapping(nodes: &[WorkerNode]) -> HashMap<ParallelUnitId, WorkerNode> {
     let mut pu_to_worker = HashMap::new();
