+++ src/frontend/src/scheduler/worker_node_manager.rs
@@ -17,7 +17,7 @@ use std::sync::{Arc, RwLock};
 
 use rand::seq::SliceRandom;
 use risingwave_common::bail;
-use risingwave_common::types::{ParallelUnitId, VnodeMapping};
+use risingwave_common::hash::vnode::{ParallelUnitId, VnodeMapping};
 use risingwave_common::util::worker_util::get_pu_to_worker_mapping;
 use risingwave_pb::common::WorkerNode;
