+++ src/storage/hummock_sdk/src/filter_key_extractor.rs
@@ -19,7 +19,7 @@ use std::time::Duration;
 
 use parking_lot::RwLock;
 use risingwave_common::catalog::ColumnDesc;
-use risingwave_common::types::VirtualNode;
+use risingwave_common::hash::vnode::VirtualNode;
 use risingwave_common::util::ordered::OrderedRowSerde;
 use risingwave_common::util::sort_util::OrderType;
 use risingwave_pb::catalog::Table;
@@ -363,12 +363,13 @@ mod tests {
     use risingwave_common::config::constant::hummock::PROPERTIES_RETENTION_SECOND_KEY;
     use risingwave_common::row::Row;
     use risingwave_common::types::ScalarImpl::{self};
-    use risingwave_common::types::{DataType, VirtualNode};
+    use risingwave_common::types::DataType;
     use risingwave_common::util::ordered::OrderedRowSerde;
     use risingwave_common::util::sort_util::OrderType;
     use risingwave_pb::catalog::Table as ProstTable;
     use risingwave_pb::plan_common::{ColumnCatalog as ProstColumnCatalog, ColumnOrder};
     use tokio::task;
+    use risingwave_common::hash::vnode::VirtualNode;
 
     use super::{DummyFilterKeyExtractor, FilterKeyExtractor, SchemaFilterKeyExtractor};
     use crate::filter_key_extractor::{
