+++ src/stream/src/common/table/state_table.rs
@@ -28,7 +28,7 @@ use risingwave_common::array::{Op, StreamChunk, Vis};
 use risingwave_common::buffer::Bitmap;
 use risingwave_common::catalog::{ColumnDesc, TableId, TableOption};
 use risingwave_common::row::{self, CompactedRow, Row, Row2, RowDeserializer, RowExt};
-use risingwave_common::types::VirtualNode;
+use risingwave_common::hash::vnode::VirtualNode;
 use risingwave_common::util::epoch::EpochPair;
 use risingwave_common::util::ordered::OrderedRowSerde;
 use risingwave_common::util::sort_util::OrderType;
