+++ src/stream/src/executor/watermark_filter.rs
@@ -19,8 +19,9 @@ use futures::StreamExt;
 use futures_async_stream::try_stream;
 use itertools::Itertools;
 use risingwave_common::bail;
+use risingwave_common::hash::vnode::VirtualNode;
 use risingwave_common::row::Row;
-use risingwave_common::types::{DataType, ScalarImpl, VirtualNode};
+use risingwave_common::types::{DataType, ScalarImpl};
 use risingwave_expr::expr::expr_binary_nonnull::new_binary_expr;
 use risingwave_expr::expr::{BoxedExpression, Expression, InputRefExpression, LiteralExpression};
 use risingwave_expr::Result as ExprResult;
